Updated patches from master-patches
Note I manually modified the QuantumManager-..start-dnsmasq...patch
so that it excludes a hunk already applied in an earlier patch.
git can handle this but patch can't.  I've asked git upstream
for how to avoid this issue with format-patch generated patches,
but pending that I've modified the patch manually to delete
the Authors part.