update /etc/sysconfig/openstack-nova-novncproxy correctly
This file was not marked as "config" in novnc,
so we include the template in this package and
mark it as config(noreplace) so that any user edits
are left in place.

rpm on upgrade will maintain the file when the new package
has the file marked as config, even if it wasn't so marked
in the original package.