avoid [keystone_authtoken] config corruption in nova.conf
Icehouse now mimics what Red Hat distro config has done for ages
and puts the keystone_authtoken _config_ in the nova _config_ file,
rather than in the api-paste.ini.

However the way it did it conflicted with our modifications to nova.conf
resulting in duplicate sections.  This in turn triggered a bug in the
puppet ini editing code that resulted in lines just containint "nil"
to be inserted, which then resulted in parsing failure in puppet and
failure to install in packstack/foreman.

Resolves: rhbz#1047156