Don't force nova to use keystone v2 api
Setting the api version here forces nova to use the keystone V2 API for
authentication and validation. This was to work around a bug that was
briefly a problem a few cycles ago and has not been around for a while.

Nova is currently tested in gate with the keystone v3 api.

Change-Id: I8683a6f9e377a126ae205b23c6c0efec9ec7f7a2