Update patch set 4
Patch Set 4: Verified-2

I don't believe setting a shell for the nova user is a reasonable default to have in packaging.

If someone wants to configure live migration over ssh/rsync, that person will have to configure other things such as ssh key based authentication for the nova user. That user should be expected to configure a SSH shell as well.

This would be an insecure default for deployments using libvirt/tcp live migrations.

Patch-set: 4