Update patch set 1
Patch Set 1: Code-Review-2

We can't do this.
The qemu-kvm-ev package is not in the RDO repositories.
It is available either in the CentOS virt SIG or the OSP repositories.

Doing a require on qemu-kvm is fine. It'll work.
If a repository that has qemu-kvm-ev is available and you do "yum install qemu-kvm", it'll install qemu-kvm-ev because qemu-kvm-ev obsoletes qemu-kvm.

Patch-set: 1
Label: Code-Review=-2