Update patch set 3
Patch Set 3: Code-Review+1

Okay, had a good chat over IRC with hguemar explaining why we need to remove that file. That's because Nova builds it anyway and we don't want to ship it in the RPM until we have a proper WSGI stack for handling it. Thanks, Haïkel :-)

Patch-set: 3
Reviewer: Gerrit User 248 <248@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1