Update patch set 1
Patch Set 1:

I asked Vladik who is the original author of the serie and he told me that pyobject2 is not really used, in libosinfo there is a package object-introspecton which does the work. So it seems that we are OK here.

Patch-set: 1
Reviewer: Gerrit User 105 <105@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0