Update patch set 1
Patch Set 1:

> So you probably also want to remove in .spec the line which
 > indicates that the services are obsolete

for ec2-api - there is no reference to ec2-api in the .spec file
for objectstore - there is still an Obsoletes: openstack-nova-objectstore which could be deprecated in the Octa cycle.  But that is a different change that should be proposed to rpm-master.

Patch-set: 1
Reviewer: Gerrit User 127 <127@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0