Update patch set 1
Patch Set 1:

> So you probably also want to remove in .spec the line which
 > indicates that the services are obsoletes

This is the removal of tiles that should have been dropped back in January.  They are only shipped in the source rpm and not any of the packaged rpms.  There was an Obsoletes entry added for openstack-nova-objectstore and that can be pruned in Octa cycle, but Not for newton-rdo in my opinion.

Patch-set: 1
Reviewer: Gerrit User 127 <127@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0