Update patch set 1
Patch Set 1: Code-Review+2

This temprevert will help us to promote CI in both TripleO and Puppet, which is something very important for us at this stage of the cycle.

Though it won't solve the issue for upgrades, as I understand we need to have the cells ready in the database when upgrading from Newton to Ocata.

Anyway, this temprevert is something we need ASAP to promote CIs and let people do their work. We can't stay blocked on this topic just because Nova hasn't figure out how to make it work outside devstack.

Thanks.

Patch-set: 1
Label: Code-Review=+2