Add explicit requirement on libvirt-client
After libvirt-2.1, libvirt-client is not a dependency
for libvirt-daemon [1], so virsh will not get installed in
nova-compute hosts unless we explicitely add it as
a requirement.

This patch adds a dependency on /usr/bin/virsh to be
ready for newer versions of libvirt.

[1] http://pkgs.fedoraproject.org/cgit/rpms/libvirt.git/commit/?h=f26&id=b1ac7b57910fa4b2ca87039c3150dcb77ca576d1

Change-Id: Ie7477ac07c61d0814598a95ea6289838ddea224c
(cherry picked from commit 375074bc7475a9fb704cdda0ed14cf72778de0a7)