Update patch set 1
Patch Set 1:

> What the context of this change?


It's to rationalize the nova libvirt auth, particularly for containerization.

We I started looking at this, the approach to containerization was to run nova as root, which gives nova access to the libvirtd socket. However this causes lots of permissions issues in /var/lib/nova/instances.

I changed this to run nova-compute as the nova user but had add a hack to override the nova polkit rule on the docker host to grant the docker nova user access (uid is not the same as the host nova uid). https://github.com/openstack/puppet-tripleo/blob/master/manifests/profile/base/docker.pp#L167

After further investigation we concluded the we gain nothing by using polkit, and it would be best to just disable it and rely on unix filesystem permissions to control access. This patch disables polkit auth in tripleo and sets the appropriate permissions on the unix socket to grant only the nova gid (and root) access: https://review.openstack.org/479816.

Sven pointed out that there is already a libvirt group, and libvirt already adds a polkit rule to grant access to this group... so it would be better to simply add nova to the libvirt group (removing the need for the nova polkit rules in this spec) and modify https://review.openstack.org/479816 to grant access to the libvirt group instead of the nova group. This would be more flexible in future if any other users/services required access to the unix socket.


> If we do that we are relaying to how libvirt is packaged, right?


We already are, it's a requirement yea?

If the question is "is it safe to rely on this behavior":
The rpm spec comes from the libvirt repo. 
There is "official support" for the libvirt group polkit rule since https://github.com/libvirt/libvirt/commit/e94979e901517af9fdde358d7b7c92cc055dd50c
This was included in RHEL 7.2 https://bugzilla.redhat.com/show_bug.cgi?id=1194593.

Patch-set: 1