Install the right variant of QEMU binary based on distribution
Ensure that the right variant (among: 'qemu-kvm', 'qemu-kvm-rhev', and
'qemu-kvm-ev') of QEMU binary depending on the distribution in use.

This aims to address:

    https://bugzilla.redhat.com/show_bug.cgi?id=1392820 -- Add RPM deps
    to require install of qemu-kvm-rhev, not qemu-kvm-rhel

To quote Dan Berrangé from IRC:

    The point of my bug [1392820] was 2 fold:

    (1) first make sure RHOS never gets 'qemu-kvm' installed;
    (2) make sure we always get the rhel-7.x update that we have
        actually tested Nova against

NOTE: We have to re-evaulate the version string for the QEMU binary for
each RHOS / RHEL release -- add a note to that effect in the spec, as a
future visual reminder.

Change-Id: I5ff6fa77fec86b7e15bf88d786208af7b2ea2050