Update patch set 3
Patch Set 3: Code-Review+1

After further discussion, I think my concern about the conditionals here is out of scope for this patch. I think the conditionals are weird and probably not required here, but I'm happy that this patch is just following what's already there.

I'm obviously not -1 on a nit, but if you wanted to fix it...

Patch-set: 3
Label: Code-Review=+1