Require python-tooz >= 1.58.0
Ie293180e22d38a58fc51d8089547dd1f66ed2249 correctly introduced this
dependency but did not set a version constraint and thus would not
ensure the minimum version of 1.44.0 was met. As we ship 1.58.0 with
>=Pike we can use that version here.

I417d69b5859bbc55fc1121b68013f124cb0b88dc had already corrected this in
Queens so this change is for Pike only.

Change-Id: I085713e795b8739d6a398bf26633cbc54aa155a4