Update patch set 1
Patch Set 1:

> varargs not passed as last argument, and this fails as a syntax
 > error with python2:- https://review.opendev.org/#/c/699237/3/nova/compute/utils.py@1375

@lyarwood, @Stephen, wdyt if this can be fixed in nova?


Me no idea if that or similar patches will be backported to stable branches, if yes it will definetly cause issue with those branches as those are running with python2?
If it cannot be fixed in nova, then we need to pin nova in RDO until we move to CentOS8 completely.

Patch-set: 1