Update requires python2-oslo-messaging >= 10.2.0
When working on testing the upgrade path from
Rocky to Train the nova-conductor service didn't
start unless python2-oslo-messaging was updated
from 8.1.4 to 10.2.0

Moving on to functional testing, live migrations
did not work on compute nodes. The compute nodes
only said this [1].

The nova-conductor log said this [2].

After upgrading python2-oslo-messaging on the
compute node and restart nova-compute the live
migration worked.

[1] http://paste.openstack.org/show/791657/
[2] http://paste.openstack.org/show/791658/

Change-Id: I399ae5ee547c63f12811f2596f93cb5d4c287514