Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> 2021-09-15 11:25:39.298162 | container |   python-pypowervm >= 1.1.15
> 2021-09-15 11:25:39.298167 | container |   PyYAML >= 5.1
> 2021-09-15 11:25:39.298171 | container |   python-zvmcloudconnector >= 1.3.0
> 
> Need to mark these as optional somehow as they aren't required when using the libvirt virt driver.

Hello Lee,
All the NVR you've updated are not available in RDO at the moment. All of them (except alembic) are provided by CentOS offical repos and we tend to not override them if possible. That's why I didn't update them to match upstream requirements.
Plus the job rdopkg-reqcheck is non-voting so nevermind if it's failed, it helps us to quickly reqcheck.

Patch-set: 2