Update patch set 2
Patch Set 2: Code-Review+1

This also makes sense to me we avoid the deprecated options in favor fo the more user friendly url
although there likely shoudl be a dependcy here on the chagne that changes keystone to run on port 5000 which i do not see in teh comment message. if you have a link to that it would be nice ot add if you respin the patch. otherwise this looks good.

Patch-set: 2
Reviewer: Gerrit User 600 <600@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1