Update patch set 5
Patch Set 5:

(1 comment)

i really dont like these files being shipped in nova-dist git but if they are at a minium this need to run under python 3.

ideally these woudl be removed form nova-distgit and moved to the nova_libvirt contianer
https://github.com/openstack/tripleo-common/tree/master/container-images/tcib/base/os/nova-base/

i belive the migration target image is using the nova_compute image
https://github.com/openstack/tripleo-heat-templates/blob/master/deployment/nova/nova-migration-target-container-puppet.yaml#L164

so either this should be injected by ooo into the image and stored in tripleo-heat-templates or it should have been added to tripleo-common and the included in the image when it was built.

we shoudl not be packaging this in the nova rpm IMO.

maintaining -1 until python version is fixed.
when that is done we can discuss keeping this temporaly in nova-distgit if other are ok with it
and fiture out where it should live longterm later.

Patch-set: 5
Reviewer: Gerrit User 600 <600@270e2033-b340-4cff-9539-693957ebf0e7>