Update virtio-vga RPM dependency
CentOS9/RHEL QEMU packaging has began to align[1] its packaging layout
to match Fedora's.  This is primarily to improve user experience when
moving from Fedora to CentOS/RHEL.  A part of the impact of this change
is its effect on OSP: the 'virtio-vga' display device has moved to its
own separate package.  Nova supports 'virti-vga'.

Specifically, the "hw-display-virtio-vga.so" file used to be part of
'qemu-kvm-common' RPM, but now it has moved to its own separate package
called 'device-display-virtio-vga'.  So update Nova spec to reflect
this.

Instead of going with a package-based Requires as below:

    Requires(pre): device-display-virtio-vga

We're going with a file-based Requres:

    Requires(pre):  %{_libdir}/qemu-kvm/hw-display-virtio-vga.so

This will transparently allow DNF to handle the updates and install the
renamed package ("device-display-virtio-vga" in tihs case) -- the file
/usr/lib64/qemu-kvm/hw-display-virtio-vga.so used to part of
"qemu-kvm-common", and the same file with same path is now part of the
renamed package.

[1] https://gitlab.com/redhat/centos-stream/src/qemu-kvm/-/merge_requests/60

Thanks: Alfredo Moralejo Alonso for the discussion.

Change-Id: I85bba1c738f3581698a4698cedb3f96d1dacd0b4
Signed-off-by: Kashyap Chamarthy <kchamart@redhat.com>