migration-wrapper: Match ssh tunnel script more loosely
Currently the ssh tunnel script used by libvirt to connect to the
host is matched using pure string comparison, which works fine
but is a bit fragile in that even minor, semantically-irrelevant
changes to libvirt might break the detection logic: this has
happened in the past, when virt-ssh-helper was introduced and the
ssh tunnel script had to be modified accordingly.

Use regex matching instead. This way, we can focus on just the
relevant bits (is virt-ssh-helper or netcat being used?) and
ignore the irrelevant ones (how many spaces appear in between
tokens?).

For parts of the command line that have historically been enclosed
in quotes but are not actually required to be from the shell's
point of view (netcat binary name and libvirt connection URI),
both the unquoted and quoted versions are accepted to guarantee
compatibility with existing libvirt versions as well as upcoming
versions that might do away with the unnecessary quoting.

Note that this change doesn't in any way negatively impact
security: even if a rogue ssh client sent over a script that is
crafted specifically to execute malicious code while at the same
time being recognized as valid by one of our patterns, that would
not do what they expect because nova-migration-wrapper will still
ultimately end up running either virt-ssh-helper or nc with
carefully vetted arguments through the call to os.execlp() in
allow_command().

Change-Id: I8f577bebc07546e319f6af1246c39c147dad00f9
Signed-off-by: Andrea Bolognani <abologna@redhat.com>