openstack-nova-25.0.0-0.1.0rc1
Changelog:
- Update to 25.0.0.0rc1

Require different virtualized displays depending on arch

virtio-vga is only provided for x86_64 and ppc64le [1], so we need to do the
requirement conditional using boolean dependencies [2]. Otherwise,
openstack-nova-compute fail to install in aarch64 systems.

For aarch64, the recommended one is virtio-gpu so we are adding it as a
requirement only for aarch64.

[1] https://gitlab.com/redhat/centos-stream/rpms/qemu-kvm/-/blob/c9s/qemu-kvm.spec#L582
[2] https://rpm-software-management.github.io/rpm/manual/boolean_dependencies.html

Change-Id: Idbb08a5cc1f70d40ee6601247d8d4cda8fd21674
(cherry picked from commit b49673146961e88fee0dd69804be1595cef3876f)